feat(components): adjusted button structure
